Delete Namespace button fixed

+++ src/components/settings-page/index.js
@@ -62,7 +62,7 @@ function SettingsAction(props) {
 
 
     return(
-        <div id="workflow-actions" className="shadow-soft rounded tile fit-content" style={{ fontSize: "11pt", padding: "0" }}>
+        <div id="workflow-actions" className="" style={{ margin: "10px 10px 0px 0px"}}>
             {/* <div className="dropdown">
                 <button onClick={(e)=>{
                     // e.stopPropagation()
@@ -80,7 +80,7 @@ function SettingsAction(props) {
                 (<></>)
                 }
             </div>  */}
-            <ConfirmButton  Icon={IoTrash} IconColor={"var(--danger-color)"} OnConfirm={(ev) => {
+            <ConfirmButton ConfirmationText={"Delete Namespace Confirmation"} Icon={IoTrash} IconColor={"var(--danger-color)"} OnConfirm={(ev) => {
                             deleteNamespace()
                             ev.stopPropagation()
               }}/>
